apps/api/requirements.txt
# Requirements for Inkspire API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
prisma>=0.10.0
python-dotenv>=1.0.0
langchain>=0.0.311
openai>=1.0.0
pytest>=7.4.3
httpx>=0.25.0
python-multipart>=0.0.6
